requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
numpy
pandas
scikit-learn
pydantic>=2
pyyaml
mlflow
prometheus-client
typer
pdfminer.six
dvc
sphinx
pytest
httpx
black
ruff